Flash Messages in admin

+++ src/LoveThatFit/AdminBundle/Controller/SurveyController.php
@@ -24,10 +24,23 @@ class SurveyController extends Controller {
         $question = new SurveyQuestion();
         $form = $this->getQuestionForm($question);
         $form->bind($request);
+       if ($form->isValid()) {
         $em = $this->getDoctrine()->getManager();
         $em->persist($question);
         $em->flush();
+        $this->get('session')->setFlash('success','Survey Question has been created');
         return $this->redirect($this->generateUrl('admin_survey'));
+       }else
+       {
+         $this->get('session')->setFlash('warning','Survey Question cantnot be created!');
+       return $this->render('LoveThatFitAdminBundle:Survey:index.html.twig', array(
+                    'data' =>  $this->getQuestionsList(),
+                    'addNewForm' => $this->getAddNewQuestionForm()->createView(),
+                    'operation'=>null,
+                    'id'=>null,
+                    'count_question'=>count($this->getQuestionsList()),
+        ));
+       }
     }
     
 public function addNewAnswerAction($question_id) {
@@ -50,10 +63,24 @@ public function addNewAnswerAction($question_id) {
         $form = $this->createFormBuilder($answers)
                 ->add('answer', 'text')
                 ->getForm();
-        $form->bind($request);      
+        $form->bind($request);
+        if ($form->isValid()) {
         $em->persist($answer);
         $em->flush();
+        $this->get('session')->setFlash('success','Answer has been created');
         return $this->redirect($this->generateUrl('admin_survey'));
+        }else
+        {
+           $this->get('session')->setFlash('warning','Answer cannot be creatd');
+            return $this->render('LoveThatFitAdminBundle:Survey:index.html.twig', array(
+                    'answerForm' => $form->createView(),
+                    'question_id' => $question_id,
+                    'data' => $this->getQuestionsList(),                    
+                    'addNewForm' => $this->getAddNewQuestionForm()->createView(),
+                    'operation'=>'AddAnwser',
+                    'count_question'=>count($this->getQuestionsList()),
+        ));
+        }
     }
 
     public function editQuestionAction($question_id) {
@@ -73,9 +100,24 @@ public function addNewAnswerAction($question_id) {
         $question = $this->getquestionById($question_id);
         $form = $this->getQuestionForm($question);
         $form->bind($request);
+        if($form->isValid())
+        {
         $em->persist($question);
         $em->flush();
+        $this->get('session')->setFlash('success','Survey Question has been updated');
         return $this->redirect($this->generateUrl('admin_survey'));
+        }else
+        {   
+            $this->get('session')->setFlash('warning','Survey Question cannot be update');
+            return $this->render('LoveThatFitAdminBundle:Survey:index.html.twig', array(
+                    'editForm' => $form->createView(),
+                    'id' => $question_id,                   
+                    'data' => $this->getQuestionsList(),
+                    'addNewForm' => $this->getAddNewQuestionForm()->createView(),
+                    'operation'=>'editQuestion',
+                    'count_question'=>count($this->getQuestionsList()),
+        ));
+        }
     }
 
     public function editAnswerAction($answer_id) {       
@@ -100,24 +142,47 @@ public function addNewAnswerAction($question_id) {
                 ->add('answer', 'text',array('label'=>' '))
                 ->getForm();
         $form->bind($request);
+        if($form->isValid())
+        {
         $em->persist($answer);
         $em->flush();
+        $this->get('session')->setFlash('success','Survey Answer has been updated');
         return $this->redirect($this->generateUrl('admin_survey'));
+        }else
+        {
+            $this->get('session')->setFlash('warning','Survey Answer cannot be update');
+            return $this->render('LoveThatFitAdminBundle:Survey:index.html.twig', array(
+                    'editAnswerForm' => $form->createView(),
+                    'id' => $answer_id,                    
+                    'data' => $this->getQuestionsList(),
+                    'addNewForm' => $this->getAddNewQuestionForm()->createView(),
+                    'operation'=>'editAnswer',
+                    'count_question'=>count($this->getQuestionsList()),            
+        ));
+        }
     }
 
     public function deleteQuestionAction(Request $request, $question_id) {
         $em = $this->getDoctrine()->getManager();
         $question = $this->getquestionById($question_id);
+        if (!$question) {
+                throw $this->createNotFoundException('Unable to find Survey Question.');
+            }
         $em->remove($question);
         $em->flush();
+        $this->get('session')->setFlash('success','Survey Question has been deleted');
         return $this->redirect($this->generateUrl('admin_survey'));
     }
 
     public function deleteAnswersAction(Request $request, $answer_id) {
         $em = $this->getDoctrine()->getManager();
         $answer = $this->getAnswerById($answer_id);
+        if (!$answer) {
+                throw $this->createNotFoundException('Unable to find Answer.');
+            }
         $em->remove($answer);
         $em->flush();
+        $this->get('session')->setFlash('success','Survey Answer has been deleted');
         return $this->redirect($this->generateUrl('admin_survey'));
     }
 //---------------------------------------------------------------------------------
